Added basic idle functionality

+++ Assets/Scripts/CashManager.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class CashManager : MonoBehaviour {
-    [SerializeField] public float CurrentCash;
+    [SerializeField] public int CurrentCash;
 
 }
